[ncneofetch] use GetSystemInfo on Windows

+++ src/lib/in.c
@@ -502,11 +502,11 @@ mouse_click(inputctx* ictx, unsigned release, char follow){
     logwarn("dropping click in margins %ld/%ld\n", y, x);
     return;
   }
-  if(x >= ictx->ti->dimx - (ictx->rmargin + ictx->lmargin)){
+  if((unsigned)x >= ictx->ti->dimx - (ictx->rmargin + ictx->lmargin)){
     logwarn("dropping click in margins %ld/%ld\n", y, x);
     return;
   }
-  if(y >= ictx->ti->dimy - (ictx->bmargin + ictx->tmargin)){
+  if((unsigned)y >= ictx->ti->dimy - (ictx->bmargin + ictx->tmargin)){
     logwarn("dropping click in margins %ld/%ld\n", y, x);
     return;
   }
